Rename filterspecs to filters and ensure that they are actually registered
